Refactor: extracted Sentiment polarity code into its own module
